Fix VObjConstraint hierarchy not respecting conditions in base classes
VObjConstraint's conditions are merged in the overloaded __add__ function. However the merged filter_cons was not returned and the select_cons was not merged at all.
Now merge select_cons, with select constraints in derived class take priority over those in base classes if the constraints are on the same item; return the merged filter_cons and select_cons.